#115 add functionality for admin access, admin page, correct routing
