+++ src/components/Dashboard.tsx
@@ -1,17 +1,44 @@
-import React from "react";
+import React, { useState } from "react";
 import { Route } from "react-router";
 import { BrowserRouter, Redirect } from "react-router-dom";
 import { ChooseApplication } from "./ChooseApplication";
-import { useRecoilValue } from "recoil";
-import { currentUserState } from "../stateManagement/userAuth";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import { currentUserState, userRoleState } from "../stateManagement/userAuth";
 import DrawerBar from "./DrawerBar";
 import Home from "./Home";
 import UserProfile from "./UserProfile";
 import { ApplicationForm } from "./ApplicationForm";
 import RecivedApplications from "./RecivedApplications";
+import { useEffect } from "react";
+import firebase from "firebase";
 
 export default function Dashboard() {
   const currentUser = useRecoilValue(currentUserState);
+  const [userRole, setUserRole] = useRecoilState(userRoleState);
+
+  var db = firebase.firestore();
+
+  useEffect(() => {
+    getUserRole();
+  }, [userRole]);
+
+  async function getUserRole() {
+    if (currentUser != null) {
+      await db
+        .collection("user")
+        .doc(currentUser.uid)
+        .get()
+        .then((doc) => {
+          const data = doc?.data();
+          if (!data) {
+            console.log("no data here");
+            return null;
+          } else {
+            setUserRole(data.role);
+          }
+        });
+    }
+  }
 
   if (currentUser == null) {
     return <Redirect to="/login" />;
@@ -27,7 +54,15 @@ export default function Dashboard() {
         <Route exact path="/studentnm" component={ApplicationForm} />
         <Route exact path="/studentleker" component={ApplicationForm} />
         <Route exact path="/studentcup" component={ApplicationForm} />
-        <Route path="/submitted" component={RecivedApplications} />
+        {userRole == "admin" ? (
+          <Route
+            exact
+            path="/recivedApplications"
+            component={RecivedApplications}
+          />
+        ) : (
+          " "
+        )}
       </BrowserRouter>
     </div>
   );
